Add an example of ws-cma-es
